% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ashe_earn_cpa}
\alias{ashe_earn_cpa}
\title{ASHE Earnings Data 2021}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 105 rows and 4 columns.
}
\usage{
ashe_earn_cpa
}
\description{
Average annual earnings published by the ONS from analysis of the Annual Survey of Hours and Earnings by 4-digit SIC-2007 industry.
Data are collapsed into the 105 categories used in the ONS Supply and Use Tables using employment calculated from the Labour Force
Survey by industry as weights to calculate average annual earnings by CPA category.
}
\details{
Figures are taken for the most recent year for which data is available. As of 05/05/2021 this is the 2020 provisional estimates. The data
are obtained \href{https://www.ons.gov.uk/employmentandlabourmarket/peopleinwork/earningsandworkinghours/datasets/industry4digitsic2007ashetable16}{here}.
}
\keyword{datasets}
