% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{awe}
\alias{awe}
\title{Average Weekly Earnings (AWE) Index}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 21 rows and 2 columns.
}
\source{
\href{https://www.ons.gov.uk/employmentandlabourmarket/peopleinwork/earningsandworkinghours/timeseries/ka5h/emp}{AWE: Whole Economy Index}
}
\usage{
awe
}
\description{
An index of average weekly earnings growth from 2000 to 2020. Used to deflate the package earnings data to any other year used for the
input-output analysis. Annual figures taken from the ONS statistical release and accessed 02/11/2021.
}
\keyword{datasets}
