% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInputs.R
\name{CalcInputs}
\alias{CalcInputs}
\title{Generate IO Model Inputs from Data}
\usage{
CalcInputs(
  data = tobalciomodel::tob_alc_data,
  year = 2019,
  prop_alc_off = -0.01,
  prop_alc_on = -0.01,
  prop_tob_fm = -0.01,
  prop_tob_ryo = -0.01
)
}
\arguments{
\item{data}{Data table. Tobacco and alcohol expenditures}

\item{year}{year of data to use.}

\item{prop_alc_off}{Proportionate change in alcohol expenditure.}

\item{prop_alc_on}{Proportionate change in alcohol expenditure.}

\item{prop_tob_fm}{Proportionate change in factory-made cigarettes expenditure.}

\item{prop_tob_ryo}{Proportionate change in hand-rolled tobacco expenditure.}
}
\value{

}
\description{
Function to use aggregate data on expenditures included in the package to estimate
changes in expenditure in basic prices following assumed percentage changes in
consumption.
}
\examples{

\dontrun{

}
}
