% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanFinalDemand.R
\name{CleanFinalDemand}
\alias{CleanFinalDemand}
\title{Clean Final Demand Data}
\usage{
CleanFinalDemand(data)
}
\arguments{
\item{data}{data table containing the product level changes in final demand. Must contain the variable 'ch_demand'}
}
\description{
Take the product level changes in final demand and collapse them into a format which can be used
by 'FinalDemandVec()' to produce the vector of final demands to be used in the input-output model.
}
