% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{coicop_cpa_mapping}
\alias{coicop_cpa_mapping}
\title{Map COICOP to CPA}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 3708 rows and 4 columns.
}
\usage{
coicop_cpa_mapping
}
\description{
A data table which can be used to map expenditure distributed across 36 COICOP
categories onto the 105 CPA categories which are used in the Supply and Use tables.
}
\details{
The mapping is derived from Table 3 - Household Final Consumption Expenditure (HhFCE) -
of the Supply and Use Tables for 2018. The data can be found \href{https://www.ons.gov.uk/economy/nationalaccounts/supplyandusetables/datasets/inputoutputsupplyandusetables}{here}.
}
\keyword{datasets}
