% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_alcohol_duty}
\alias{data_alcohol_duty}
\title{Alcohol duties as of September 2020.}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 34 rows and 10 columns.
}
\usage{
data_alcohol_duty
}
\description{
Alcohol duties as of September 2020.
}
\keyword{datasets}
