% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_iotable_fai}
\alias{data_iotable_fai}
\title{Fraser of Allender Institute (FAI) Input-Output Table.}
\format{
An object of class \code{data.frame} with 106 rows and 117 columns.
}
\usage{
data_iotable_fai
}
\description{
Fraser of Allender Institute (FAI) Input-Output Table.
}
\keyword{datasets}
