% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_iotables_ons}
\alias{data_iotables_ons}
\title{Office for National Statistics (ONS) Input-Output Tables.}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 2310 rows and 117 columns.
}
\usage{
data_iotables_ons
}
\description{
Office for National Statistics (ONS) Input-Output Tables.
}
\keyword{datasets}
