% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{data_mesas_englandwales}
\alias{data_mesas_englandwales}
\title{Alcohol Consumption Data for England and Wales 2000-2019.}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 20 rows and 22 columns.
}
\usage{
data_mesas_englandwales
}
\description{
Alcohol Consumption Data for England and Wales 2000-2019.
}
\keyword{datasets}
