% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demand_changes.R
\name{demand_changes}
\alias{demand_changes}
\title{Construct the Vector of Final Demand Changes}
\usage{
demand_changes(change.off = NULL, change.on = NULL, change.tob = NULL)
}
\arguments{
\item{change.off}{total change in consumption of off-trade alcohol (£m)}

\item{change.on}{total change in consumption of on-trade alcohol (£m)}

\item{change.tob}{total change in consumption of tobacco (£m)}
}
\description{
Based on the scenario selected by `tobalciomodel::scenarios`, or manual input, construct a vector of changes
in final demand of length 106 and input the changes in off-trade alcohol, on-trade alcohol, and tobacco consumption
to be modelled.
}
