% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{demo_data}
\alias{demo_data}
\title{Demo Data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 3 rows and 18 columns.
}
\usage{
demo_data
}
\description{
Synthetic data used to represent a simplified 3-sector economy for demonstration of the methodology used in this package.
}
\keyword{datasets}
