% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EconEffectsCalc.R
\name{EconEffectsCalc}
\alias{EconEffectsCalc}
\title{Calculate Economic Impacts}
\usage{
EconEffectsCalc(
  leontief,
  fdemand,
  FAI = FALSE,
  year = 2019,
  tax_data = tobalciomodel::inctax_params
)
}
\arguments{
\item{leontief}{List object. The output of \code{LeonfiefCalc} containing multipliers and leontief matrices.}

\item{fdemand}{Data table. The output of \code{PrepFinalDemand}.}

\item{FAI}{Logical. If TRUE, uses the Fraser of Allender Institute (FAI) table instead of the
ONS ones. Defaults to FALSE.}

\item{year}{Integer. Select the year (from 2016 to 2020) of alcohol, tobacco, earnings and employment data to use in the analysis.}

\item{tax_data}{Data table. Package data containing the necessary parameters to calculate
income tax and national insurance contributions from annual earnings data}
}
\value{
A data table of economic impacts by sector
}
\description{
Apply the model parameters to the change in final demand vector in order to
model the impact of changes in demand on output, gross value added, and
employment.
}
\examples{

\dontrun{


}
}
