% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elasticities}
\alias{elasticities}
\title{Alcohol own-price elasticities.}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 10 rows and 2 columns.
}
\usage{
elasticities
}
\description{
Price elasticity of demand for 10 alcohol products, taken from Meng et al. (2014)
}
\keyword{datasets}
