% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elasticities_alc}
\alias{elasticities_alc}
\title{Price elasticities for Alcohol}
\format{A data frame with 10 observations and 4 variables.
\describe{
    \item{Product}{Beer, Cider, Wine, Spirits, RTDs}
    \item{On.Trade}{On-trade alcohol price elasticities}
    \item{Off.Trade}{Off-trade alcohol price elasticities }
    \item{Source}{Source for the elasticities. Collis et al. (2010) or Meng et al. (2014)}
}}
\usage{
elasticities_alc
}
\description{
A data frame containing the price elasticity of demand for five types of alcohol; beer, cider,
wine, spirits, and RTDs split by off and on trade sales. Elasticities are obtained from multiple sources.
}
\keyword{datasets}
