% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{elasticities_tob}
\alias{elasticities_tob}
\title{Price elasticities for Tobacco}
\format{A data frame with 10 observations and 4 variables.
\describe{
    \item{product}{cigarettes (need to add RYO vs FM to the data)}
    \item{Elasticity}{price elasticities}
    \item{Source}{Source for the elasticities}
}}
\usage{
elasticities_tob
}
\description{
A data frame containing the price elasticity of demand for five types of alcohol; beer, cider,
wine, spirits, and RTDs split by off and on trade sales. Elasticities are obtained from multiple sources.
}
\keyword{datasets}
