% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{employment}
\alias{employment}
\title{LFS Employment Data for 106 Sectors.}
\format{
A data frame with 106 observations and 23 variables.
}
\usage{
employment
}
\description{
Total employment generated from the quarterly Labour Force Survey (LFS) by 4-digit SIC-2007 industry. Data
is then collapsed into the 106 sectors (including 3 alcohol disaggregated sectors) used in the FAI IO table.
Employment for the disaggregated sectors is split from the parent sector proportionately by total sectoral
output measured in 2010 in the IO table.
}
\details{
For each year from 2010 to 2020, total employment and total employment measured in full-time equivalents (FTE)
is available (individuals who report being part-time are considered 0.5 FTE)
}
\keyword{datasets}
