% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmplTaxEffects.R
\name{EmplTaxEffects}
\alias{EmplTaxEffects}
\title{Calculate Employment Tax Effects}
\usage{
EmplTaxEffects(effects, earnings_data = tobalciomodel::ashe_earn_cpa)
}
\arguments{
\item{effects}{Data table. Containing type 0 and type 1 effects on employment
by sector. Output of the \code{EconEffectsCalc} function.}

\item{earnings_data}{Data table. Average annual earnings by sector.}
}
\value{

}
\description{
Uses estimates of the change in employment by sector and data on average annual salaries
by sector to calculate estimates of the change in tax revenues from employment - income
tax, employee national insurance contributions (NICs), and employer NICs.
}
\examples{
\dontrun{

### calculate economic effects

econ_effect <- EconEffectsCalc(leontief,
                           final_demand_vec)

### calculate the tax effects

taxes <- EmplTaxEffects(effects = econ_effect)

}
}
