% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/final_demand_inputs_tob.R
\name{final_demand_inputs_tob}
\alias{final_demand_inputs_tob}
\title{Preparation to Model Change in Tobacco Final Demand}
\usage{
final_demand_inputs_tob(yr = 2010, elasticity = "meng14")
}
\arguments{
\item{yr}{which years data to use?.}

\item{elasticity}{meng14 or collis10 - the source used for price elasticities.}

\item{scotland}{default is FALSE - indicates whether to use Scottish data over the default England/Wales.}
}
\value{
A dataset containing price, elasticity, expenditure, and consumption information split
        by on and off trade for each of the five alcohol products. Consumption and expenditure measured in billions.
}
\description{
This function filters the package data by chosen input parameters and merges the elasticity and MESAS information
into a single dataset ready to be used to analyse the effects of policy changes on final demand.
}
