% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gdhi}
\alias{gdhi}
\title{Gross Disposable Household Income}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 120 rows and 3 columns.
}
\usage{
gdhi
}
\description{
Stores GDHI data from 1997 - 2020 for the UK and the four constituent nations. Figures are presented in £millions
and in basic prices. These data are used for the denominator for household expenditure in type 2 technical
coefficient matrices
}
\details{
\source \href{https://www.ons.gov.uk/economy/regionalaccounts/grossdisposablehouseholdincome/datasets/regionalgrossdisposablehouseholdincomegdhi}{Office for National Statistics}.
}
\keyword{datasets}
