% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInputs.R
\name{GenInputs}
\alias{GenInputs}
\title{Generate IO Model Inputs from Data}
\usage{
GenInputs(
  data = tobalciomodel::tob_alc_data,
  prop_alc_on = -0.01,
  prop_alc_off = -0.01,
  prop_tob_fm = -0.01,
  prop_tob_ryo = -0.01,
  govt = TRUE,
  yr = 2020
)
}
\arguments{
\item{data}{Data table. Tobacco and alcohol expenditures}

\item{prop_alc_on}{Proportionate change in alcohol expenditure.}

\item{prop_alc_off}{Proportionate change in alcohol expenditure.}

\item{prop_tob_fm}{Proportionate change in factory-made cigarettes expenditure.}

\item{prop_tob_ryo}{Proportionate change in hand-rolled tobacco expenditure.}

\item{govt}{Logical. If TRUE include changes in government revenue in the analysis.}

\item{yr}{year of data to use.}
}
\value{

}
\description{
Function to use aggregate data on expenditures included in the package to estimate
changes in expenditure in basic prices following assumed percentage changes in
consumption.
}
\examples{

\dontrun{

}
}
