% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenInputsAGGDATA.R
\name{GenInputsAGGDATA}
\alias{GenInputsAGGDATA}
\title{Generate IO Inputs from Aggregate Data}
\usage{
GenInputsAGGDATA(scenario_list, year = 2019, input_labels, write_inputs, name)
}
\arguments{
\item{scenario_list}{List object. List of numeric vectors of length four of proportionate changes to demand to model. In order;
off-trade alcohol, on-trade alcohol, factory-made cigarettes, and hand-rolled tobacco.}

\item{year}{Numeric. Year of consumption and economic data to use}

\item{input_labels}{Character vector - vector of labels identifying the scenarios modelled.}

\item{write_inputs}{Character - file path to the folder to which the input data will be written.}

\item{name}{Character - file name for the input data.}
}
\value{

}
\description{
Calculate changes in demand based on aggregate spending data for alcohol and tobacco for
specified years. Input proportionate changes to demand to create a data table of changes
in final demand.
}
\examples{

\dontrun{

}
}
