% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hhold_reallocate.R
\name{hhold_reallocate}
\alias{hhold_reallocate}
\title{Allow Households to Adjust Expenditure}
\usage{
hhold_reallocate(data = NULL, pro.rata = TRUE)
}
\arguments{
\item{data}{dataset created by `final_demand_inputs()`}

\item{pro.rata}{if TRUE, consumption is reallocated according to the existing distribution of
household spending. If FALSE, manually select sectors to re-allocate to.}
}
\value{
A vector of changes in final demand measured in £bn
}
\description{
Using the vector of changes in final demand computed using the  `final_demand_inputs()` function, allow households
to keep their real consumption constant by increasing/decreasing consumption on goods not directly modelled.
}
