% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_calculate.R
\name{impact_calculate}
\alias{impact_calculate}
\title{Calculate Economic Impacts}
\usage{
impact_calculate(demand.change = NULL, multipliers = NULL)
}
\arguments{
\item{demand.change}{list object created by `simulate_alcohol_policy()`}
}
\value{
A table of model outputs
}
\description{
Take the vector of changes in final demand and use the relevant multipliers to calculate the
output, employment, and gross value added impacts of the policy.
}
