% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ImpactCalc.R
\name{ImpactCalc}
\alias{ImpactCalc}
\title{Calculate Economic Impacts}
\usage{
ImpactCalc(
  demand.change = fdemand_fai_govt$final_demand,
  multipliers = leontief_fai,
  select_year = 2010,
  FAI = TRUE,
  export = FALSE,
  path = "output/",
  name = "example"
)
}
\arguments{
\item{demand.change}{vector of changes in final demand}

\item{multipliers}{list object created by `LeontiefCalc()`.}

\item{select_year}{year of analysis.}

\item{FAI}{logical. TRUE if using the Fraser of Allender Institute IO table (the default). Select FALSE to use one of the ONS tables.}

\item{export}{logical; if TRUE saves a csv file of results.}

\item{path}{filepath for exporting results.}

\item{name}{name for results file.}
}
\value{
A table of model outputs
}
\description{
Take the vector of changes in final demand and use the relevant multipliers to calculate the
output, employment, and gross value added impacts of the policy.
}
