% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{inctax_params}
\alias{inctax_params}
\title{Income Tax Parameters}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 11 rows and 13 columns.
}
\source{
\href{https://www.gov.uk/government/collections/tax-structure-and-parameters-statistics}{Tax structure and parameters statistics}
}
\usage{
inctax_params
}
\description{
Data table of income tax and national insurance parameters - thresholds and tax rates - used for calculating tax paid on employment incomes
based on annual earnings.
}
\keyword{datasets}
