% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_analysis.R
\name{io.analysis}
\alias{io.analysis}
\title{Wrapper Function to Produce a Full IO Analysis}
\usage{
io.analysis(
  scenario.num = NULL,
  hhold = FALSE,
  hhold.pro.rata = TRUE,
  hhold.sectors = NULL,
  hhold.prop = NULL
)
}
\arguments{
\item{hhold}{allow households to re-allocate spending}

\item{hhold.pro.rata}{arguments to pass to `hhold_reallocate()` - pro rata redistribute}

\item{hhold.sectors}{arguments to pass to `hhold_reallocate()` - specific sectors to re-allocate to}

\item{hhold.prop}{arguments to pass to `hhold_reallocate()` - proportions to re-allocate by sector}

\item{scenario}{the row number of the data frame `tobalciomodel::scenarios` corresponding to the chosen scenario to model.}
}
\description{
Run all functions to model a policy scenario and produce output of an IO analysis, for a given
pre-loaded scenario option.
}
