% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOModelDemo.R
\name{IOModelDemo}
\alias{IOModelDemo}
\title{Input-Output Model Example}
\usage{
IOModelDemo(data = tobalciomodel::demo_data, fdemand_vec = c(0, -1000, 0))
}
\arguments{
\item{data}{data table. Package data containing the parameters for the synthetic economy.}

\item{fdemand_vec}{Numeric vector. A vector of changes in final demand. Must be length 3.}
}
\value{

}
\description{
Runs input-output analyses using a simplified 3-sector economy.
}
