% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iotable}
\alias{iotable}
\title{Input-Output Table.}
\format{
A data frame with 106 observations and 116 variables.
\describe{
    \item{name}{Name of the SIC07-based sector of economic activity.}
    \item{sec1-sec106}{For each sector, a variable capturing inter-industry sales/purchases.}
    \item{hhold.demand}{The vector of household demand (a component of final demand).}
    \item{final.demand}{The vector of final demands (all non-intermediate demand).}
    \item{hhold.output}{The vector of employee compensation - the value added to total output by household labour supply.}
    \item{total.output}{The vector of total output - value of raw materials, plus imports, returns to capital and labour, taxes.}
    \item{total.demand}{Should equal total output.}
    \item{gva.taxes}{gross value added - taxes less subsidies on production component}
    \item{gva.wages}{gross value added - employee compensation component}
    \item{gva.gos}{gross value added - gross operating surplus component}
    \item{gva.total}{gross value added total - sum of employee compensation, gross operating surplus, and taxes on production}
    }
}
\usage{
iotable
}
\description{
A data frame containing relevant data extracted from the publicly available
alcohol-disaggregated IO table used by the Fraser of Allender Institute.
}
\keyword{datasets}
