% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iotable_nire}
\alias{iotable_nire}
\title{Input-Output Tables - Northern Ireland}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 126 rows and 76 columns.
}
\usage{
iotable_nire
}
\description{
Input-output flow-table as well as demand and output vectors derived from data tables provided by the
\href{https://www.gov.scot/publications/input-output-latest/}{Northern Ireland Statistics and Research Authority
(NISRA)} for the years 2016 and 2017. More information about the
tables is available \href{https://www.nisra.gov.uk/statistics/economic-accounts-project/analytical-input-output-tables}{here}.
}
\keyword{datasets}
