% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iotable_ons}
\alias{iotable_ons}
\title{Input-Output Table}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 2310 rows and 123 columns.
}
\usage{
iotable_ons
}
\description{
Input-output flow-table as well as demand and output vectors derived from the ONS supply and use tables
1997 - 2018. In addition to the same variables obtained from the FAI input-output table, additional information
is available on purchaser prices, tax as a proportion of the purchaser price, and industry profit margins.
}
\details{
Supply and Use Tables are published by the Office for National Statistics. The raw data tables from which this
dataset is derived can be found
\href{https://www.ons.gov.uk/economy/nationalaccounts/supplyandusetables/datasets/inputoutputsupplyandusetables}{here}.
}
\keyword{datasets}
