% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iotable_scot}
\alias{iotable_scot}
\title{Input-Output Tables - Scotland}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 98 rows and 110 columns.
}
\usage{
iotable_scot
}
\description{
Input-output flow-table as well as demand and output vectors derived from data tables provided by the
\href{https://www.gov.scot/publications/input-output-latest/}{Scottish Government} for the years
1998, 2010 2015, and 2019. More information about the tables is available
\href{https://www.gov.scot/publications/about-supply-use-input-output-tables/pages/user-guide-symmetric-tables/}{here}.
}
\keyword{datasets}
