% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lfs_empl}
\alias{lfs_empl}
\title{LFS Employment Data 2010-2020}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 2937 rows and 6 columns.
}
\usage{
lfs_empl
}
\description{
Total employment is calculated from the quarterly Labour Force Survey (LFS) by 4-digit SIC-2007 industry. Data
are currently available for years 2010-2020 and in the formats required for the UK (Fraser of Allender Institute),
Scotland (Scottish government), and Northern Ireland (Northern Ireland Statistics and Research Agency) input-output tables.
}
\details{
Figures are available for total employment and total full-time equivalent (FTE)
employment (employees who report being part-time in the LFS are treated as 0.5 FTE).

The employment figures are constructed from the LFS micro-data obtained from the UK Data Service
and processed using the \href{https://stapm.gitlab.io/r-packages/lfsclean/}{`lfsclean` package v1.0.0}.
}
\keyword{datasets}
