% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lfs_empl_cpa}
\alias{lfs_empl_cpa}
\title{LFS Employment Data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 1155 rows and 5 columns.
}
\usage{
lfs_empl_cpa
}
\description{
Total employment is calculated from the quarterly Labour Force Survey (LFS) by 4-digit SIC-2007 industry. Data
are collapsed into the 105 CPA categories used in the ONS Supply and Use Tables and for a given year, the average total
employment across the four quarters is taken.
}
\details{
Figures are available for total employment and total full-time equivalent (FTE)
employment (employees categorised in the LFS as part-time are considered 0.5 FTE).

The employment figures are constructed from the LFS micro-data obtained from the UK Data Service
and processed using the `lfsclean` package v0.9.0.
}
\keyword{datasets}
