% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{macro}
\alias{macro}
\title{Macroeconomic Data.}
\format{
A data frame with 10 observations and 5 variables.
\describe{
    \item{year}{}
    \item{cpih_index}{inflation index (base year 2015)}
    \item{gdp}{nominal Gross Domestic Product (£millions)}
    \item{gva}{nominal Gross Value Added (£millions)}
    \item{emp}{aggregate employment (full-time equivalents)}
    \item{gdp_real}{Real Gross Domestic Product in 2010 prices (£millions)}
    \item{gva_real}{Real Gross Value Added in 2010 prices (£millions)}
}
}
\usage{
macro
}
\description{
Total employment generated from the quarterly Labour Force Survey (LFS) by 4-digit SIC-2007 industry. Data
is then collapsed into the 106 sectors (including 3 alcohol disaggregated sectors) used in the FAI IO table.
Employment for the disaggregated sectors is split from the parent sector proportionately by total sectoral
output measured in 2010 in the IO table.
}
\details{
For each year from 2010 to 2019, total employment and total employment measured in full-time equivalents (FTE)
is available (individuals who report being part-time are considered 0.5 FTE)
}
\keyword{datasets}
