% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{macro_data}
\alias{macro_data}
\title{Macroeconomic Data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 100 rows and 6 columns.
}
\source{
\href{https://www.ons.gov.uk/economy/nationalaccounts/supplyandusetables/datasets/inputoutputsupplyandusetablessummarytables}{supply and use summary tables}.
}
\usage{
macro_data
}
\description{
Employment data from the Labour Force Survey. Total gross value added (GVA) and
total output in £m obtained from the supply and use summary tables.
}
\details{
All values are totalled by broad industry grouping (10). These data are used to
scale calculated economic effects to the size of the economy.
}
\keyword{datasets}
