% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mesas}
\alias{mesas}
\title{Alcohol Consumption Data for England, Wales, and Scotland 2000-2019.}
\format{
A data frame with 320 observations and 12 variables.
\describe{
    \item{year}{Demands of Sector A from the row sector.}
    \item{product}{Total alcohol, and 7 types of alcohol; spirits, RTDs, fortifies wines, wine, cider, perry, and beer.}
    \item{litres.ontrade}{total volume of on-trade alcohol sales (1000 litres)}
    \item{litres.offtrade}{total volume of off-trade alcohol sales (1000 litres)}
    \item{units.pp.ontrade}{total volume of on-trade alcohol sales (units) per adult (aged 16+ years)}
    \item{units.pp.offtrade}{total volume of off-trade alcohol sales (units) per adult (aged 16+ years)}
    \item{price.ontrade}{average price per unit of alcohol sold through the on-trade.}
    \item{price.offtrade}{average price per unit of alcohol sold through the off-trade.}
    \item{population}{total adult population mid-year.}
    \item{consumption.ontrade}{total estimated on-trade consumption - units per person X price per unit X population.}
    \item{consumption.offtrade}{total estimated off-trade consumption - units per person X price per unit X population.}
    \item{country}{indicator. England and Wales, or Scotland}
}
}
\usage{
mesas
}
\description{
A data frame containing alcohol consumption data separately for Scotland, and for England and Wales.
Data was compiled by Monitoring and Evaluating Scotlands Alcohol Strategy (MESAS) from Nielsen and CGA Strategy.
Off-trade alcohol sales have been adjusted to account for the exclusion of discount retailers.
Raw data was obtained from \href{http://www.healthscotland.scot/publications/mesas-monitoring-report-2020}{the MESAS monitoring report 2020.}
}
\keyword{datasets}
