% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mesas_eng_wales}
\alias{mesas_eng_wales}
\title{MESAS England and Wales}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 50 rows and 12 columns.
}
\source{
\href{https://www.publichealthscotland.scot/publications/mesas-monitoring-report-2021/}{MESAS monitoring report 2021}.
}
\usage{
mesas_eng_wales
}
\description{
Data on alcohol consumption and prices published by MESAS - Monitoring and Evaluating Scotland's Alcohol Strategy - for
England and Wales. Data are for 2016-2020.
}
\keyword{datasets}
