% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelIO.R
\name{ModelIO}
\alias{ModelIO}
\title{Wrapper Function to Produce a Full IO Analysis}
\usage{
ModelIO(
  select_year = 2019,
  ons_table_year = 2019,
  FAI = TRUE,
  pref_changes = FALSE,
  tax_policy = TRUE,
  reallocate = TRUE,
  alcohol_data = tobalciomodel::data_mesas_englandwales,
  elasticities_data = tobalciomodel::elasticities,
  duties_data = tobalciomodel::data_alcohol_duty,
  off_trade_change = c(0, 0, 0, 0, 0),
  on_trade_change = c(0, 0, 0, 0, 0),
  beer_duty = c(1, 1, 1),
  cider_duty = c(0, 0, 0, 0, 0),
  wine_duty = c(0, 0, 0, 0, 0, 0, 0),
  spirit_duty = c(0),
  rtd_duty = c(0)
)
}
\arguments{
\item{select_year}{year of price, consumption, and employment data to use.}

\item{ons_table_year}{if using the ONS IO tables, select which years table to use}

\item{FAI}{logical. TRUE if using the Fraser of Allender Institute alcohol-disaggregated IO table. If FALSE, uses the ONS tables.}

\item{pref_changes}{logical. If TRUE, model a change in consumer preferences.}

\item{tax_policy}{logical. If TRUE, model a change in alcohol duties.}

\item{reallocate}{logical. If TRUE apply the reallocation functions to the initially calculated final demand vector.}

\item{alcohol_data}{MESAS aggregate alcohol data.}

\item{elasticities_data}{alcohol price elasticities.}

\item{duties_data}{alcohol duties as of January 2021.}

\item{off_trade_change}{vector of length 5, proportionate change in consumption of off-trade beer, cider, wine, spirits, and RTDs due to preference changes.}

\item{on_trade_change}{vector of length 5, proportionate change in consumption of on-trade beer, cider, wine, spirits, and RTDs due to preference changes.}
}
\description{
Run all functions to model a policy scenario and produce output of an IO analysis, for a given
pre-loaded scenario option. Not that only one of `pref_changes` OR `tax_policy` should be specified
as TRUE for a given analysis run.
}
