% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrefChanges.R
\name{PrefChanges}
\alias{PrefChanges}
\title{Model a Change in Final Demand due to Preferences}
\usage{
PrefChanges(
  alcohol_data = tobalciomodel::data_mesas_englandwales,
  off_trade_change = c(-0.1, -0.1, -0.1, -0.1, -0.1),
  on_trade_change = c(-0.1, -0.1, -0.1, -0.1, -0.1),
  select_year = 2019
)
}
\description{
Apply an exogenous change in the demand for alcohol.
}
