% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepFinalDemand.R
\name{PrepFinalDemand}
\alias{PrepFinalDemand}
\title{Setup Final Demand Vector}
\usage{
PrepFinalDemand(
  hhold_exp,
  govt_exp,
  hhold_saving = 0.1,
  hhold_vector = "hhfce_noalctob",
  govt_vector = "central"
)
}
\arguments{
\item{hhold_exp}{Numeric. Change in household consumption measured in basic prices.}

\item{govt_exp}{Numeric. Change in government spending.}

\item{hhold_saving}{Numeric. Assumed household savings rate.}

\item{hhold_vector}{Character. How household spending is redistributed.}

\item{govt_vector}{Character. How government spending is redistributed.}
}
\value{

}
\description{
Input changes in household and government expenditures, specify assumptions
about the distributions of spending and the household saving rate, and produce
the resulting vector of changes in final demand to model.
}
