% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessOutputs.R
\name{ProcessOutputs}
\alias{ProcessOutputs}
\title{Generate Model Outputs}
\usage{
ProcessOutputs(data, sort = NULL, FAI = FALSE)
}
\arguments{
\item{data}{data table. Output of the input-output modelling}

\item{sort}{character. Variable on which to sort sectors by effect size. Inputs are one from
c("output","empl","gva","inctaxes","earn")}

\item{FAI}{Logical. If TRUE, uses the Fraser of Allender Institute (FAI) table. If FALSE, uses the
ONS supply and use tables. Defaults to FALSE}
}
\value{

}
\description{
Take the output of the model and generate summary outputs
}
\examples{

\dontrun{

}
}
