% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadInputs.R
\name{ReadInputs}
\alias{ReadInputs}
\title{Read IO Model Inputs from STAPM Output}
\usage{
ReadInputs(inputs)
}
\arguments{
\item{inputs}{Data table. Output from the `stapmr` package function `EconCalc()`}
}
\value{

}
\description{
Function to read in and further process the outputs of STAPM price policy
modelling economic outcomes. Produces the change in demand for tobacco, on-trade
alcohol and off-trade alcohol measured in basic prices and also the net change
in government revenue.
}
\examples{

\dontrun{

### Read in the output of a price policy simulation

### Obtain economic outcomes


### use the output from EconCalc as the input to the
### function to clean IO model inputs.

iomodel_inputs <- ReadInputs(econ_calcs)

}
}
