% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadIOAT.R
\name{ReadIOAT}
\alias{ReadIOAT}
\title{Read in Input-Output Tables}
\usage{
ReadIOAT(country, year_ioat = 2010, fte = TRUE)
}
\arguments{
\item{country}{Character. Country of analysis. Options are c("UK","Scotland","N_Ireland").}

\item{year_ioat}{Numeric. Year of the input-output analytical tables used.}

\item{fte}{Logical. If TRUE (default) use full-time equivalent (FTE) employment, if FALSE
use total employees.}
}
\description{
Read and process the input-output analytical table (IOAT) for the UK economy.
}
\details{
Supply and Use Tables are published by the Office for National Statistics. The tables which this
function is designed to use are downloaded from
\href{https://www.ons.gov.uk/economy/nationalaccounts/supplyandusetables/datasets/inputoutputsupplyandusetables}{this ONS page}.
}
