% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadSUT.R
\name{ReadSUT}
\alias{ReadSUT}
\title{Read in Supply and Use Tables}
\usage{
ReadSUT(year_sut = 2018, fte = TRUE, FAI = FALSE)
}
\arguments{
\item{year_sut}{Integer. Select the year (from 1997 to 2018) ONS supply and use tables to use in the analysis.}

\item{fte}{Logical. If TRUE (default) use full-time equivalent (FTE) employment, if FALSE
use total employees.}

\item{FAI}{Logical. If TRUE, uses the Fraser of Allender Institute (FAI) table instead of the
ONS ones. Defaults to FALSE.}
}
\description{
Read and process the Supply and Use Tables (SUTs) for the UK economy which cover the period 1997-2018.
}
\details{
Supply and Use Tables are published by the Office for National Statistics. The tables which this
function is designed to use are downloaded from
\href{https://www.ons.gov.uk/economy/nationalaccounts/supplyandusetables/datasets/inputoutputsupplyandusetables}{this ONS page}.
}
