% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReallocateGovt.R
\name{ReallocateGovt}
\alias{ReallocateGovt}
\title{Reallocate Government Expenditure}
\usage{
ReallocateGovt(
  expenditure = 10,
  saving_rate = 0,
  vector = 1,
  vectors_data = tobalciomodel::vectors_govt,
  FAI = FALSE
)
}
\arguments{
\item{expenditure}{change in consumption, measured in basic prices.}

\item{vector}{Numeric (1-5). The distribution of reallocation of spending to implement from the \code{vectors_govt} data.
Option 1 (default) allocates pro-rata according to the distribution of total government spending,
option 2 allocates according to central government spending only, option 3 allocates according
to local government only. Option 4 allocates all spending to health, and option 5 allocates all spending
to education.}

\item{FAI}{Logical. If TRUE, uses the Fraser of Allender Institute (FAI) table instead of the
ONS ones. Defaults to FALSE.}

\item{govt_saving}{Numeric. Assumed government savings rate.}

\item{vector_data}{data table containing the redistribution vectors.}
}
\description{
A function to allocate increased government expenditure achieved through a duties policy
across the 105 CPA sectors.
}
