% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReallocateHhold.R
\name{ReallocateHhold}
\alias{ReallocateHhold}
\title{Reallocate Household Expenditure}
\usage{
ReallocateHhold(
  expenditure = c(-20, 10, 30),
  saving_rate = 0.1,
  vector = "hhfce_noalctob",
  vectors_data = tobalciomodel::vectors_hhold,
  mapping = tobalciomodel::coicop_cpa_mapping,
  FAI = FALSE
)
}
\arguments{
\item{expenditure}{Numeric vector. Change in household consumption measured in basic prices for off-trade alcohol,
on-trade alcohol, and tobacco.}

\item{saving_rate}{proportion of saved expenditure that will be saved rather than redistributed.}

\item{vector}{character. The distribution of reallocation of spending to implement.
Valid options are the column names of the \code{vectors_hhold} data. The default is to allocate spending
pro-rata according to the 2018 distribution of consumer spending with alcohol and tobacco excluded.}

\item{mapping}{data table containing the mapping algorithm from COICOP to CPA.}

\item{FAI}{Logical. If TRUE, uses the Fraser of Allender Institute (FAI) table instead of the
ONS ones. Defaults to FALSE.}

\item{vector_data}{data table containing the redistribution vectors.}
}
\description{
A function to reallocate, based on assumptions about how individuals
redistribute spending, the expenditure saved as a result of a policy
or exogenous changes in consumer preferences.
}
