% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scenarios}
\alias{scenarios}
\title{Analysis Scenarios}
\format{A data frame with 10 observations and 4 variables.
\describe{
    \item{year}{year of data to use}
    \item{scotland}{TRUE if applying scenario to Scotland}
    \item{elasticity}{source of elasticity information}
    \item{emp.measure}{measure of employment - fte, employment, or hours}
    \item{alc.policy}{alcohol policy scenario - exog, MUP, tax, or NULL for no policy}
    \item{tob.policy}{tobacco policy scenario - exog, MUP, tax, or NULL for no policy}
}}
\usage{
scenarios
}
\description{
A data frame containing pre-specified input parameters for scenario analyses.
}
\keyword{datasets}
