% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_scenario.R
\name{select_scenario}
\alias{select_scenario}
\title{Select Scenario Parameters}
\usage{
select_scenario(num.scenario = NULL)
}
\arguments{
\item{num.scenario}{the row number of the data frame `tobalciomodel::scenarios` corresponding to the chosen scenario to model.}
}
\description{
Choose the scenario from the pre-loaded data frame of scenarios in `tobalciomodel::scenarios`
and extract the relevant parameters to put into the IO model.
}
