% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sic_cpa_fai_mapping}
\alias{sic_cpa_fai_mapping}
\title{Mapping Products and Sectors}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 612 rows and 6 columns.
}
\usage{
sic_cpa_fai_mapping
}
\description{
Allows mapping between product CPA categories and sector IOC categories by linking both classifications
to the standard industrial classification (SIC) 2007.
}
\keyword{datasets}
