% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_tobacco_policy.R
\name{simulate_tobacco_policy}
\alias{simulate_tobacco_policy}
\title{Model Change in Final Demand due to Tobacco Policy}
\usage{
simulate_tobacco_policy(data = NULL, tob.policy = "exog", cigs.ch = 0)
}
\arguments{
\item{data}{dataset created by `final_demand_inputs()`}

\item{tob.policy}{alcohol policy type to be modelled - c("exog","MUP","tax")}

\item{cigs.ch}{a vector of proportionate consumption changes to be modelled}
}
\value{
A vector of changes in final demand measured in £bn
}
\description{
Using selected data from the `final_demand_inputs()` function, model the policy-induced change in
final demand for tobacco consumption.
}
