% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaxPol.R
\name{TaxPol}
\alias{TaxPol}
\title{Model a Change in Final Demand due to Alcohol Duty}
\usage{
TaxPol(
  alcohol_data = tobalciomodel::data_mesas_englandwales,
  elasticities_data = tobalciomodel::data_elasticities,
  duties_data = tobalciomodel::data_alcohol_duty,
  beer_duty = c(1, 1, 1),
  cider_duty = c(1, 1, 1, 1, 1),
  wine_duty = c(1, 1, 1, 1, 1, 1, 1),
  spirit_duty = c(1),
  rtd_duty = c(1),
  select_year = 2019,
  allow_cross = "No"
)
}
\description{
Apply a change in the duties on alcohol and calculate the effect on price, consumption,
and total change in demand for each alcohol product.
}
