% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tob_alc_data}
\alias{tob_alc_data}
\title{Tobacco and Alcohol Data 2010-2020}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 110 rows and 14 columns.
}
\source{
\href{https://www.publichealthscotland.scot/publications/mesas-monitoring-report-2021/}{MESAS monitoring report 2021}.

\href{https://www.gov.uk/government/statistics/alcohol-bulletin}{Alcohol Bulletin}

\href{https://www.gov.uk/government/statistics/tobacco-bulletin}{Tobacco Bulletin}.

\href{https://www.ons.gov.uk/economy/inflationandpriceindices/timeseries/d7cb/mm23}{ONE Consumer Price Index (RPI) time series: All Tobacco}

\href{https://www.ons.gov.uk/economy/inflationandpriceindices/timeseries/czmp}{ONS Average Price - Cigarettes pack of 20}
}
\usage{
tob_alc_data
}
\description{
Data on tobacco expenditure, consumption and prices obtained from the ONS and HMRC duty receipts and clearance data available in the
ONS Tobacco Bulletin and other ONS sources. Data on alcohol expenditures, consumption and prices constructed using
price and consumption data published by MESAS - Monitoring and Evaluating Scotland's Alcohol Strategy. MESAS data are combined with
data on total duty receipts by HMRC published in the ONS Alcohol Bulletin.
}
\details{
The key variables in the data are `exp_bp` and `exp_bp_2020`. These are the estimates of total expenditure in basic prices (market prices
net of tax). `exp_bp_2020` applies 2020 prices throughout and is therefore a measure of real expenditures.
}
\keyword{datasets}
