% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tobacco_data}
\alias{tobacco_data}
\title{Tobacco Data}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 10 rows and 5 columns.
}
\usage{
tobacco_data
}
\description{
Data on tobacco consumption and prices obtained from ONS and other sources.
}
\keyword{datasets}
