% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vectors_govt}
\alias{vectors_govt}
\title{Government Redistribution Vectors}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 105 rows and 10 columns.
}
\usage{
vectors_govt
}
\description{
A dataset of vectors which can be used to distribute government spending across the 105
CPA categories. These include pro-rata distributions for central government spending, local
government spending, and total public sector spending. Other vectors allow allocation of all
spending to public admin, defence, and compulsory social security; education services; health services;
caring and social work; cultural services including libraries, museums, archives.
}
\keyword{datasets}
