% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vectors_hhold}
\alias{vectors_hhold}
\title{Household Redistribution Vectors}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 36 rows and 4 columns.
}
\usage{
vectors_hhold
}
\description{
A dataset of vectors which can be used to distribute expenditure of households
across 36 COICOP categories of consumer spending. These include pro-rata distributions
of household consumption across products according to the 2018 ONS Supply and Use Tables.
THese can optionally be specified to include or exclude tobacco and/or alcohol.
}
\keyword{datasets}
