% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteResultsTab.R
\name{WriteResultsTab}
\alias{WriteResultsTab}
\title{Write Results}
\usage{
WriteResultsTab(path, template, label, out_label, col_label, write_output)
}
\arguments{
\item{path}{Character. File path to where the outputs of InputOutput are stored.}

\item{template}{Character. File path to where the results spreadsheet is stored.}

\item{label}{Character. Label for the batch of STAPM model runs.}

\item{out_label}{Character vector. Vector of IO model results RDS files suffixes (after "iomodel_output").}

\item{col_label}{Character vector. Must be the same length as out_label. Label for each IO model
to write to the columns of the spreadsheet.}

\item{write_output}{Character. File path to the folder in which to save the spreadsheet.}
}
\value{
Excel table of results
}
\description{
Pass a set of results from InputOutput to this function to write out a comparison spreadsheet of results
based on the excel template iomodel_comparison_sheet.xlsx and save to a specified output folder.
}
\examples{

\dontrun{

}
}
