% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{alcohol_data}
\alias{alcohol_data}
\title{Alcohol Data 2010-2020}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 88 rows and 15 columns.
}
\source{
\href{https://www.publichealthscotland.scot/publications/mesas-monitoring-report-2021/}{MESAS monitoring report 2021}.

\href{https://www.gov.uk/government/statistics/alcohol-bulletin}{Alcohol Bulletin}
}
\usage{
alcohol_data
}
\description{
Data on alcohol expenditures, consumption and prices. Data are constructed using price and consumption data published by MESAS - Monitoring
and Evaluating Scotland's Alcohol Strategy. Expenditures at basic prices are constructed by combining with data on total duty receipts by HMRC
published in the ONS alcohol bulletin.
}
\keyword{datasets}
