% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EconEffectsCalc.R
\name{EconEffectsCalc}
\alias{EconEffectsCalc}
\title{Calculate Economic Impacts}
\usage{
EconEffectsCalc(leontief, final_demand_vec)
}
\arguments{
\item{leontief}{List. The output of \code{LeonfiefCalc} containing multipliers and leontief matrices.}

\item{final_demand_vec}{Data table. The output of \code{PrepFinalDemand}.}
}
\value{
A data table of economic impacts by sector
}
\description{
Apply the model parameters to the change in final demand vector in order to
model the impact of changes in demand on output, gross value added, and
employment.
}
\examples{

\dontrun {

### construct the final demand vector

final_demand_vec <- tobalciomodel::PrepFinalDemand(hhold_exp = -20,
                                                   govt_exp = 10,
                                                   hhold_saving = 0.1,
                                                   hhold_vector = "hhfce_noalctob",
                                                   govt_vector = "central")

### derive the leontief inverse and multipliers

tables <- tobalciomodel::ReadSUT(path = "inputs",
                                 year = 2018,
                                 fte = TRUE)

leontief <- tobalciomodel::LeontiefCalc(tables)

### calculate economic effects

effects <- tobalciomodel::EconEffectsCalc(leontief,
                                          final_demand_vec)

}
}
