% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FinalDemandVector.R
\name{FinalDemandVec}
\alias{FinalDemandVec}
\title{Construct the Vector of Final Demand Changes}
\usage{
FinalDemandVec(data = NULL, FAI = TRUE)
}
\arguments{
\item{data}{vector of changes in final demand for; off-trade alcohol, on-trade alcohol, tobacco (£m)}

\item{FAI}{logical. TRUE if using the Fraser of Allender Institute IO table (the default). Select FALSE to use one of the ONS tables.}
}
\description{
Based on the calculated changes in final demand, construct a vector of final demand changes with the
previous calculations as inputs. Calibrate the final demand vector to the input-output table chosen
for the analysis.
}
