% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetInputs.R
\name{GenInputs}
\alias{GenInputs}
\title{Generate IO Model Inputs from Data}
\usage{
GenInputs(
  tobacco = tobalciomodel::tobacco_data,
  alcohol_eng = tobalciomodel::mesas_eng_wales,
  alcohol_scot = tobalciomodel::mesas_scotland,
  prop_alc_on = -0.1,
  prop_alc_off = -0.1,
  prop_tob_fm = -0.1,
  prop_tob_ryo = -0.1,
  yr = 2020
)
}
\arguments{
\item{tobacco}{Data table. Tobacco expenditures}

\item{alcohol_eng}{Data table. Alcohol expenditures data for England and Wales.}

\item{alcohol_scot}{Data table. Alcohol expenditures data for Scotland.}

\item{prop_alc_on}{Proportionate change in alcohol expenditure.}

\item{prop_alc_off}{Proportionate change in alcohol expenditure.}

\item{prop_tob_fm}{Proportionate change in factory-made cigarettes expenditure.}

\item{prop_tob_ryo}{Proportionate change in hand-rolled tobacco expenditure.}

\item{yr}{year of data to use.}
}
\value{

}
\description{
Function to use aggregate data on expenditures included in the package to estimate
changes in expenditure in basic prices following assumed percentage changes in
consumption.
}
