% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenInputsSTAPM.R
\name{GenInputsSTAPM}
\alias{GenInputsSTAPM}
\title{Generate IO Inputs from STAPM Models}
\usage{
GenInputsSTAPM(
  root = NULL,
  label = NULL,
  label2 = NULL,
  min_age = NULL,
  index_year = NULL,
  policy_effect_year = NULL,
  n_years = 1,
  write_inputs,
  name
)
}
\arguments{
\item{root}{Character - the file path to the folder storing STAPM raw output}

\item{label}{Character - the label used to identify the set of model runs.}

\item{label2}{Character vector - vector of treatment labels identifying specific model runs.}

\item{min_age}{Numeric - Youngest age included in the model.}

\item{index_year}{Numeric - Model index year.}

\item{policy_effect_year}{Numeric - Year in which new price policy is implemented in STAPM.}

\item{n_years}{Numeric - Number of years including the policy effect year for which to generate inputs
(defaults to 1 - policy effect year only).}

\item{write_inputs}{Character - file path to the folder to which the input data will be written.}

\item{name}{Character - file name for the input data.}
}
\value{
Data
}
\description{
Function to read in and further process the outputs of STAPM price policy
modelling economic outcomes. Produces the change in demand for tobacco, on-trade
alcohol and off-trade alcohol measured in basic prices and also the net change
in government revenue.
}
\examples{

\dontrun{

}
}
