% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_calculate.R
\name{impact_calculate}
\alias{impact_calculate}
\title{Calculate Economic Impacts}
\usage{
impact_calculate(
  demand.change = NULL,
  multipliers = NULL,
  export = FALSE,
  path = "output/"
)
}
\arguments{
\item{demand.change}{list object created by `simulate_alcohol_policy()`}

\item{multipliers}{data frame of multipliers used to calculate the impacts.}

\item{export}{logical; if TRUE saves a csv file of results.}

\item{path}{filepath for exporting results.}
}
\value{
A table of model outputs
}
\description{
Take the vector of changes in final demand and use the relevant multipliers to calculate the
output, employment, and gross value added impacts of the policy.
}
