% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_calculate.R
\name{impact_calculate}
\alias{impact_calculate}
\title{Calculate Economic Impacts}
\usage{
impact_calculate(
  demand.change = NULL,
  multipliers = NULL,
  yr = NULL,
  base = 2010,
  export = FALSE,
  path = "output/",
  name = "example"
)
}
\arguments{
\item{demand.change}{vector of changes in final demand created by `demand_changes()`}

\item{multipliers}{data frame of multipliers used to calculate the impacts.}

\item{yr}{year of analysis.}

\item{base}{base year for real terms comparisons.}

\item{export}{logical; if TRUE saves a csv file of results.}

\item{path}{filepath for exporting results.}

\item{name}{name for results file.}
}
\value{
A table of model outputs
}
\description{
Take the vector of changes in final demand and use the relevant multipliers to calculate the
output, employment, and gross value added impacts of the policy.
}
