% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_analysis.R
\name{io.analysis}
\alias{io.analysis}
\title{Wrapper Function to Produce a Full IO Analysis}
\usage{
io.analysis(scenario = NULL)
}
\arguments{
\item{scenario}{the row number of the data frame `tobalciomodel::scenarios` corresponding to the chosen scenario to model.}
}
\description{
Run all functions to model a policy scenario and produce output of an IO analysis, for a given
pre-loaded scenario option.
}
