% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IOModel.R
\name{IOModel}
\alias{IOModel}
\title{Input-Output Wrapper Function}
\usage{
IOModel(
  FAI = FALSE,
  year_sut = 2018,
  year = 2019,
  fte = TRUE,
  hhold_exp,
  govt_exp,
  hhold_saving = 0,
  hhold_reallocate = "hhfce_noalctob",
  govt_reallocate = "central",
  tax_data = tobalciomodel::inctax_params
)
}
\arguments{
\item{FAI}{Logical. If TRUE, uses the Fraser of Allender Institute (FAI) table. If FALSE, uses the
ONS supply and use tables. Defaults to FALSE}

\item{year_sut}{Integer. Select the year (from 1997 to 2018) ONS supply and use tables to use in the analysis.}

\item{year}{Integer. Select the year (from 2016 to 2020) of alcohol, tobacco, earnings and employment data to use in the analysis.}

\item{fte}{Logical. If TRUE (default) use full-time equivalent (FTE) employment, if FALSE
use total employees.}

\item{hhold_exp}{Numeric vector of length 3. Change in household consumption measured in basic prices for - in order - off-trade alcohol,
on-trade alcohol, and tobacco.}

\item{govt_exp}{Numeric. Change in government spending.}

\item{hhold_saving}{Numeric. Assumed household savings rate.}

\item{hhold_reallocate}{Character. Name of the vector which determines how household spending is redistributed.}

\item{govt_reallocate}{Character. Name of the vector which determines how government spending is redistributed.}

\item{tax_data}{Data table. Package data containing the necessary parameters to calculate
income tax and national insurance contributions from annual earnings data}
}
\value{

}
\description{
Read in inputs, make calculations, and return clean model outputs for an input-output analysis. Wrapper
function for the other functions in the package to implement the full workflow. This function takes parameter
inputs determined externally (e.g. from processed outputs obtained via the Sheffield Tobacco and Alcohol Policy Model
- STAPM).
}
\examples{

\dontrun{


}
}
\author{
Damon Morris
}
