% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{iotable_fai}
\alias{iotable_fai}
\title{Input-Output Table}
\format{
An object of class \code{data.frame} with 106 rows and 117 columns.
}
\usage{
iotable_fai
}
\description{
Input-output flow-table as well as demand and output vectors from the alcohol-disaggregated
input-output table constructed by Fraser of Allender Institute (FAI).
}
\keyword{datasets}
