% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeontiefCalc.R
\name{LeontiefCalc}
\alias{LeontiefCalc}
\title{Calculate L Matrices and Multipliers}
\usage{
LeontiefCalc(list)
}
\arguments{
\item{list}{List. The output of the ReadSUT() function containing the IO table, supply table, and
technical coefficients.}
}
\description{
Take the processed Input-Output and Supply and Use Tables and derive the Leontief inverse matrix and
multipliers required for the economic impact analysis.
}
