% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeontiefCalc.R
\name{LeontiefCalc}
\alias{LeontiefCalc}
\title{Calculate L Matrices and Multipliers}
\usage{
LeontiefCalc(list, country, year_ioat)
}
\arguments{
\item{list}{List. The output of the ReadIOAT() function containing the IO table and
technical coefficients.}

\item{country}{Character. Country of analysis. Options are c("UK","scotland","nireland").}

\item{year_ioat}{Numeric. Year of the input-output analytical tables used.}
}
\description{
Take the processed Input-Output and Supply and Use Tables and derive the Leontief inverse matrix and
multipliers required for the economic impact analysis. The function produces multipliers and Leontief
matrices for type 0 (direct), type 1 (indirect), and type 2 (induced) effects on output, GVA, and employment.
}
