% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LeontiefCalc.R
\name{LeontiefCalc}
\alias{LeontiefCalc}
\title{Calculate L Matrices and Multipliers}
\usage{
LeontiefCalc(list, FAI = TRUE)
}
\arguments{
\item{list}{List. The output of the ReadSUT() function containing the IO table, supply table, and
technical coefficients.}

\item{FAI}{Logical. If TRUE, uses the Fraser of Allender Institute (FAI) table instead of the
ONS ones. Defaults to FALSE.}
}
\description{
Take the processed Input-Output and Supply and Use Tables and derive the Leontief inverse matrix and
multipliers required for the economic impact analysis. The function produces multipliers and Leontief
matrices for type 0 (direct), type 1 (indirect), and type 2 (induced) effects on output, GVA, and employment/
}
