% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multipliers.R
\name{multipliers}
\alias{multipliers}
\title{Produce Sectoral Output Multipliers}
\usage{
multipliers()
}
\description{
Calculate the Leontief inverse Type I and Type II matrices from the package data and calculate
the sector by sector multipliers for output, GVA, and employment.
}
