% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepFinalDemand.R
\name{PrepFinalDemand}
\alias{PrepFinalDemand}
\title{Setup Final Demand Vector}
\usage{
PrepFinalDemand(
  hhold_exp,
  govt_exp,
  hhold_saving = 0,
  govt_saving = 0,
  hhold_reallocate = 3,
  govt_reallocate = 1,
  FAI = FALSE
)
}
\arguments{
\item{hhold_exp}{Numeric vector of length 3. Change in household consumption measured in basic prices for off-trade alcohol,
on-trade alcohol, and tobacco.}

\item{govt_exp}{Numeric. Change in government spending.}

\item{hhold_saving}{Numeric. Assumed household savings rate.}

\item{govt_saving}{Numeric. Assumed government savings rate.}

\item{hhold_reallocate}{Numeric (1-3). The distribution of reallocation of spending to implement from the \code{vectors_hhold} data.
Option 1 allocates pro-rata across all consumption categories, option 2 excludes alcohol and tobacco consumption, option 3
(default) further excludes health, education, rents and utilities.}

\item{govt_reallocate}{Numeric (1-5). The distribution of reallocation of spending to implement from the \code{vectors_govt} data.
Option 1 (default) allocates pro-rata according to the distribution of total government spending,
option 2 allocates according to central government spending only, option 3 allocates according
to local government only. Option 4 allocates all spending to health, and option 5 allocates all spending
to education.}

\item{FAI}{Logical. If TRUE, uses the Fraser of Allender Institute (FAI) table instead of the
ONS ones. Defaults to FALSE.}
}
\value{

}
\description{
Input changes in household and government expenditures, specify assumptions
about the distributions of spending and the household saving rate, and produce
the resulting vector of changes in final demand to model.
}
\examples{

\dontrun{

}
}
