% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrepFinalDemand.R
\name{PrepFinalDemand}
\alias{PrepFinalDemand}
\title{Setup Final Demand Vector}
\usage{
PrepFinalDemand(
  hhold_exp,
  govt_exp,
  hhold_saving = 0.1,
  hhold_vector = "hhfce_noalctob",
  govt_vector = "central",
  FAI = FALSE
)
}
\arguments{
\item{hhold_exp}{Numeric vector. Change in household consumption measured in basic prices for off-trade alcohol,
on-trade alcohol, and tobacco.}

\item{govt_exp}{Numeric. Change in government spending.}

\item{hhold_saving}{Numeric. Assumed household savings rate.}

\item{hhold_vector}{Character. How household spending is redistributed.}

\item{govt_vector}{Character. How government spending is redistributed.}

\item{FAI}{Logical. If TRUE, uses the Fraser of Allender Institute (FAI) table instead of the
ONS ones. Defaults to FALSE.}
}
\value{

}
\description{
Input changes in household and government expenditures, specify assumptions
about the distributions of spending and the household saving rate, and produce
the resulting vector of changes in final demand to model.
}
\examples{

\dontrun{

## construct a final demand vector for a fall in household spending in basic
## prices of £20m and an increase in government spending of £10m.

## Assume households and governments distribute spending pro-rata as per
## household final consumption expenditure and central government spending
## distributions measured in 2018.

## Households assumed to save 10\% of freed-up spending.

finaldemand <- tobalciomodel::PrepFinalDemand(hhold_exp = -20,
govt_exp = 10,
hhold_saving = 0.1,
hhold_vector = "hhfce_noalctob",
govt_vector = "central")

}
}
