% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProcessOutputs.R
\name{ProcessOutputs}
\alias{ProcessOutputs}
\title{Generate Model Outputs}
\usage{
ProcessOutputs(data, sort)
}
\arguments{
\item{data}{data table. Output of the input-output modelling}

\item{sort}{character. Variable on which to sort sectors by effect size. Inputs are one from
c("output","empl","gva","inctaxes","earn")}
}
\value{

}
\description{
Take the output of the model and generate summary outputs
}
\examples{

\dontrun{

}
}
