% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadInputs.R
\name{ReadInputs}
\alias{ReadInputs}
\title{Read IO Model Inputs from STAPM Output}
\usage{
ReadInputs(data, policy_effect_year, n_years = 1, govt = TRUE)
}
\arguments{
\item{data}{Data table. Output from the `stapmr` package function `EconCalc()`}

\item{policy_effect_year}{Numeric. Year for which economic impacts are to be modelled.}

\item{n_years}{Numeric. Number of years starting with
the policy effect year for which to extract treatment effects.}

\item{govt}{Logical. If TRUE include changes in government revenue in the analysis.}
}
\value{
Data
}
\description{
Function to process the outputs of STAPM price policy
modelling economic outcomes. Produces the change in demand for tobacco, on-trade
alcohol and off-trade alcohol measured in basic prices and also the net change
in government revenue.
}
\examples{

\dontrun{

### Read in the output of a price policy simulation

### Obtain economic outcomes


### use the output from EconCalc as the input to the
### function to clean IO model inputs.

iomodel_inputs <- ReadInputs(econ_calcs)

}
}
