% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadInputs.R
\name{ReadInputs}
\alias{ReadInputs}
\title{Read IO Model Inputs from STAPM Output}
\usage{
ReadInputs(
  inputs,
  policy_effect_year,
  quantity = FALSE,
  quantity_prop = c(-0.1, -0.1, -0.1)
)
}
\arguments{
\item{inputs}{Data table. Output from the `stapmr` package function `EconCalc()`}

\item{policy_effect_year}{Numeric. Year for which economic impacts are to be modelled.}

\item{quantity}{Logical. If TRUE calculate the effects of a non-price induced change in quantity consumed.
Changes are calculated relative to the control arm of the simulation. Defaults to FALSE.}

\item{quantity_prop}{Numeric vector of length 3. The proportionate exogenous change in consumption to model for
off-trade alcohol, on-trade alcohol, and tobacco respectively.}
}
\value{

}
\description{
Function to read in and further process the outputs of STAPM price policy
modelling economic outcomes. Produces the change in demand for tobacco, on-trade
alcohol and off-trade alcohol measured in basic prices and also the net change
in government revenue.
}
\examples{

\dontrun{

### Read in the output of a price policy simulation

### Obtain economic outcomes


### use the output from EconCalc as the input to the
### function to clean IO model inputs.

iomodel_inputs <- ReadInputs(econ_calcs)

}
}
