% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReallocateGovt.R
\name{ReallocateGovt}
\alias{ReallocateGovt}
\title{Reallocate Government Expenditure}
\usage{
ReallocateGovt(
  expenditure = 10,
  vector = "central",
  vectors_data = tobalciomodel::vectors_govt
)
}
\arguments{
\item{expenditure}{change in consumption, measured in basic prices.}

\item{vector}{character. The distribution of reallocation of spending to implement.
Valid options are the column names of the \code{vectors_hhold} data. The default is to
allocate spending pro-rata according to the 2018 distribution of central government spending.}

\item{vector_data}{data table containing the redistribution vectors.}
}
\description{
A function to allocate increased government expenditure achieved through a duties policy
across the 105 CPA sectors.
}
