% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReallocateGovt.R
\name{ReallocateGovt}
\alias{ReallocateGovt}
\title{Reallocate Government Expenditure}
\usage{
ReallocateGovt(
  revenue,
  govt_passthru = 0,
  vector = 1,
  vectors_data = tobalciomodel::vectors_govt,
  FAI = FALSE
)
}
\arguments{
\item{revenue}{Numeric. change in government revenues, measured in basic prices.}

\item{govt_passthru}{Numeric. Assumed government rate of passthrough - the proportion of change in revenues
which are matched by changes in government expenditure. Defaults to 0 (no passthrough).}

\item{vector}{Numeric (1-5). The distribution of reallocation of spending to implement from the \code{vectors_govt} data.
Option 1 (default) allocates pro-rata according to the distribution of total government spending,
option 2 allocates according to central government spending only, option 3 allocates according
to local government only. Option 4 allocates all spending to health, and option 5 allocates all spending
to education.}

\item{FAI}{Logical. If TRUE, uses the Fraser of Allender Institute (FAI) table instead of the
ONS ones. Defaults to FALSE.}

\item{vector_data}{data table containing the redistribution vectors.}
}
\description{
A function to allocate increased government expenditure achieved through a duties policy
across the 105 CPA sectors.
}
\examples{

\dontrun{

### generate a vector of government final demand changes for a fall in
### revenues of £100m. With passthrough = 0, this will produce a vector
### of 0s.

govt_spend <- ReallocateGovt(revenue = -100,
                             govt_passthru = 0,
                             vector = 1,
                             vectors_data = tobalciomodel::vectors_govt,
                             FAI = FALSE)

### generate a vector of government final demand changes for a fall in
### revenues of £100m. With passthrough = 1, with vector = 1, this will
### impose a total cut in government spending of £100m, allocated
### pro-rata across the distribution of total govt spending.

govt_spend <- ReallocateGovt(revenue = -100,
                             govt_passthru = 1,
                             vector = 1,
                             vectors_data = tobalciomodel::vectors_govt,
                             FAI = FALSE)

### Passthrough = 0.5 - for a £100m fall in revenues, a cut in spending of $50m,
### vector = 2 - spending cuts allocated across the distribution of central
###              government spending

govt_spend <- ReallocateGovt(revenue = -100,
                             govt_passthru = 0.5,
                             vector = 2,
                             vectors_data = tobalciomodel::vectors_govt,
                             FAI = FALSE)

}
}
