% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReallocateGovt.R
\name{ReallocateGovt}
\alias{ReallocateGovt}
\title{Reallocate Government Revenues}
\usage{
ReallocateGovt(
  data = NULL,
  final_demand = NULL,
  FAI = TRUE,
  select_year = 2010
)
}
\arguments{
\item{data}{the data table of duties, prices, and consumption produced by the `TaxPol()` function}

\item{final_demand}{the vector of final demand changes calulcated by the `FinalDemandVec()` function.}

\item{FAI}{logical. TRUE if using the Fraser of Allender Institute IO table (the default). Select FALSE to use one of the ONS tables.}

\item{select_year}{year of data to use if using the ONS tables.}
}
\description{
Allocate net revenues as a result of a tax policy change to other sectors, reflecting an increase
in government spending due to an increase in revenues. Adjust the change in final demand vector to
reflect this reallocation before conducting input-output analysis.
}
