% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReallocateHHold.R
\name{ReallocateHHold}
\alias{ReallocateHHold}
\title{Reallocate Household Spending}
\usage{
ReallocateHHold(
  final_demand = NULL,
  FAI = TRUE,
  select_year = 2010,
  prorata = TRUE,
  sectors = NULL,
  prop = NULL
)
}
\arguments{
\item{final_demand}{the vector of final demand changes calculated by the `FinalDemandVec()` function.}

\item{FAI}{logical. TRUE if using the Fraser of Allender Institute IO table (the default). Select FALSE to use one of the ONS tables.}

\item{select_year}{year of data to use if using the ONS tables.}
}
\description{
Allocate household consumption as a result of modelled changes in preferences.
}
