% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReallocateHhold.R
\name{ReallocateHhold}
\alias{ReallocateHhold}
\title{Reallocate Household Expenditure}
\usage{
ReallocateHhold(
  expenditure = -20,
  saving_rate = 0.1,
  vector = "hhfce_noalctob",
  vectors_data = tobalciomodel::vectors_hhold,
  mapping = tobalciomodel::coicop_cpa_mapping
)
}
\description{
Reallocate Household Expenditure
}
