% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_alcohol_policy.R
\name{simulate_alcohol_policy}
\alias{simulate_alcohol_policy}
\title{Model Change in Final Demand due to Alcohol Policy}
\usage{
simulate_alcohol_policy(
  data = NULL,
  alc.policy = "exog",
  on.trade.ch = c(0, 0, 0, 0, 0),
  off.trade.ch = c(0, 0, 0, 0, 0),
  reallocate = FALSE,
  prob = FALSE
)
}
\arguments{
\item{data}{dataset created by `final_demand_inputs()`}

\item{alc.policy}{alcohol policy type to be modelled - c("exog","MUP","tax")}

\item{on.trade.ch}{a vector of proportionate consumption changes to be modelled for Beer, Cider, RTDs, Spirits, Wine}

\item{off.trade.ch}{a vector of proportionate consumption changes to be modelled for Beer, Cider, RTDs, Spirits, Wine}

\item{reallocate}{logical; if TRUE then saved expenditures are reallocated to other sectors.}
}
\value{
A vector of changes in final demand measured in £bn
}
\description{
Using selected data from the `final_demand_inputs()` function, model the policy-induced change in
final demand for alcohol and tobacco consumption.
}
