% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tobacco_data}
\alias{tobacco_data}
\title{Tobacco Price and Consumption Data}
\format{
A data frame with 10 observations and 4 variables.
\describe{
    \item{year}{year of data to use}
    \item{population}{source of elasticity information}
    \item{expenditure}{measure of employment - fte, employment, or hours}
    \item{prevalance}{alcohol policy scenario - exog, MUP, tax, or NULL for no policy}
    \item{ncigs}{average number of cigs smoked per day}
    \item{price}{RRP of a pack of 20 cigarettes}
    \item{price.per.cig}{price per cigarette}
    \item{consumption}{expenditure divided by price}
}
}
\usage{
tobacco_data
}
\description{
A data frame containing expenditure, quantity, and price time series data
for tobacco consumption.
}
\keyword{datasets}
